Added file existence check for config.js

+++ install.sh
@@ -67,7 +67,7 @@ setup_app_config() {
 		echo "- Importing existing data..."
 		echo "  configs...";
 		cp ${APP_HOME}-${INSTALL_TIME}/etc/metrilyx/metrilyx.conf ${APP_HOME}/etc/metrilyx/metrilyx.conf;
-		cp ${APP_HOME}-${INSTALL_TIME}/metrilyx/static/config.js ${APP_HOME}/metrilyx/static/config.js;
+		[[ -f "${APP_HOME}-${INSTALL_TIME}/metrilyx/static/config.js" ]] || cp ${APP_HOME}-${INSTALL_TIME}/metrilyx/static/config.js ${APP_HOME}/metrilyx/static/config.js;
 		echo "  dashboards..."
 		cp -a ${APP_HOME}-${INSTALL_TIME}/pagemodels/ ${APP_HOME}/pagemodels/;
 		echo "  heatmap index..."
